Migrate rest of feature tests over
