+++ fusion/tests/src/Concerns/InstallsFusion.php
@@ -50,6 +50,8 @@ trait InstallsFusion
 
         Artisan::call('fusion:sync');
         Artisan::call('fusion:flush');
+
+        activity()->enableLogging();
     }
 
     /**
