+++ fusion/tests/stubs/MockModule/module.json
