+++ fusion/tests/stubs/MockModule/src/Http/Resources/AcmeResource.php
