+++ fusion/tests/stubs/MockModule/src/Providers/ModuleServiceProvider.php
