--- tests/Foundation/DatabaseMigrations.php
@@ -1,25 +1,0 @@
-<?php
-
-
-namespace Tests\Foundation;
-
-use Illuminate\Contracts\Console\Kernel;
-
-trait DatabaseMigrations
-{
-    /**
-     * Define hooks to migrate the database before and after each test.
-     *
-     * @return void
-     */
-    public function runDatabaseMigrations()
-    {
-        $this->artisan('module:migrate');
-
-        $this->app[Kernel::class]->setArtisan(null);
-
-        $this->beforeApplicationDestroyed(function () {
-            $this->artisan('module:migrate:rollback');
-        });
-    }
-}
