--- tests/Foundation/Traits/CreatesApplication.php
@@ -1,21 +1,0 @@
-<?php
-
-
-namespace Tests\Foundation\Traits;
-
-trait CreatesApplication
-{
-    /**
-     * Create the application.
-     *
-     * @return \Illuminate\Foundation\Application
-     */
-    public function createApplication()
-    {
-        $app = require __DIR__ . '/../../../bootstrap/app.php';
-
-        $app->make(\Illuminate\Contracts\Console\Kernel::class)->bootstrap();
-
-        return $app;
-    }
-}
