--- tests/Foundation/Traits/WithFactories.php
@@ -1,23 +1,0 @@
-<?php
-
-
-namespace Tests\Foundation\Traits;
-
-use Illuminate\Database\Eloquent\Factory as ModelFactory;
-
-trait WithFactories
-{
-    /**
-     * Load model factories from path.
-     *
-     * @param  string  $path
-     *
-     * @return $this
-     */
-    protected function withFactories($path)
-    {
-        $this->app->make(ModelFactory::class)->load($path);
-
-        return $this;
-    }
-}
