--- tests/Stubs/Importer/Entries.csv
@@ -1,5 +1,0 @@
-id,name,slug,summary,description,status,location
-1,Post A,post-a,This is a summary for post a.,"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",1,"123 Main St Seattle, WA 98123"
-2,Post B,post-b,This is a summary for post b.,Sed enim ut sem viverra aliquet. Vitae auctor eu augue ut lectus. Urna porttitor rhoncus dolor purus non enim praesent. Ridiculus mus mauris vitae ultricies leo integer malesuada nunc. Integer enim neque volutpat ac tincidunt vitae. Nec ultrices dui sapien eget mi proin sed libero. Dictumst vestibulum rhoncus est pellentesque elit. At consectetur lorem donec massa sapien faucibus et molestie. Ut diam quam nulla porttitor massa id neque aliquam vestibulum. Sed ullamcorper morbi tincidunt ornare massa eget egestas. Consequat nisl vel pretium lectus quam id. Bibendum enim facilisis gravida neque convallis a cras semper auctor.,0,
-,Post C,post-c,This is a summary for post c.,Nullam vehicula ipsum a arcu cursus vitae. Viverra suspendisse potenti nullam ac tortor vitae purus faucibus. Porttitor rhoncus dolor purus non enim praesent elementum. Posuere sollicitudin aliquam ultrices sagittis orci a scelerisque purus. Vitae purus faucibus ornare suspendisse sed nisi lacus sed. Elit at imperdiet dui accumsan sit amet nulla. Amet porttitor eget dolor morbi non. Sit amet volutpat consequat mauris nunc. Integer malesuada nunc vel risus commodo viverra maecenas accumsan lacus. Ac turpis egestas maecenas pharetra convallis posuere morbi. Amet volutpat consequat mauris nunc. Volutpat lacus laoreet non curabitur gravida. Hac habitasse platea dictumst quisque. Mauris in aliquam sem fringilla ut morbi tincidunt augue interdum.,1,"321 Wall St Seattle, WA 98321"
-,Post D,post-d,This is a summary for post d.,Sed viverra ipsum nunc aliquet. Natoque penatibus et magnis dis parturient. Elit pellentesque habitant morbi tristique senectus et. Diam in arcu cursus euismod quis viverra. Egestas tellus rutrum tellus pellentesque. Purus ut faucibus pulvinar elementum integer enim neque. Ipsum faucibus vitae aliquet nec ullamcorper sit amet risus. Etiam dignissim diam quis enim lobortis scelerisque fermentum. Mauris nunc congue nisi vitae suscipit tellus. Quis lectus nulla at volutpat diam ut.,1,
